auth/app/requirements.txt
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
pycparser==2.21
PyJWT==2.5.0
pymongo==4.2.0
python-dotenv==0.21.0
Werkzeug==2.2.2
